requirements.txt
openai
colorama
tiktoken
matplotlib
python-dotenv
streamlit
pandas
